Rename [Alice|Bob]Spawn traits to Spawn[Alice|Bob]
