+++ cnd/src/swap_protocols/rfc003/alice/mod.rs
@@ -1,9 +1,9 @@
 mod actions;
 mod initiate_request;
 mod send_request;
-mod spawner;
+mod spawn_alice;
 
-pub use self::{actions::*, initiate_request::*, send_request::*, spawner::*};
+pub use self::{actions::*, initiate_request::*, send_request::*, spawn_alice::*};
 
 use crate::swap_protocols::{
     asset::Asset,
