+++ cnd/src/swap_protocols/rfc003/alice/spawn_alice.rs
@@ -4,11 +4,10 @@ use crate::{
         asset::Asset,
         rfc003::{
             self,
-            bob::{self, State, SwapCommunication},
-            create_ledger_events::CreateLedgerEvents,
+            alice::{self, State, SwapCommunication},
             state_machine,
             state_store::StateStore,
-            Ledger,
+            CreateLedgerEvents, Ledger,
         },
         LedgerConnectors,
     },
@@ -20,9 +19,9 @@ use futures_core::{
 };
 use std::sync::Arc;
 
-pub trait BobSpawn: Send + Sync + 'static {
+pub trait SpawnAlice: Send + Sync + 'static {
     #[allow(clippy::type_complexity)]
-    fn bob_spawn<AL: Ledger, BL: Ledger, AA: Asset, BA: Asset>(
+    fn spawn_alice<AL: Ledger, BL: Ledger, AA: Asset, BA: Asset>(
         &self,
         swap_request: rfc003::Request<AL, BL, AA, BA>,
         response: rfc003::Response<AL, BL>,
@@ -30,12 +29,12 @@ pub trait BobSpawn: Send + Sync + 'static {
         LedgerConnectors: CreateLedgerEvents<AL, AA> + CreateLedgerEvents<BL, BA>;
 }
 
-impl<S> BobSpawn for Connector<S>
+impl<S> SpawnAlice for Connector<S>
 where
     S: Send + Sync + 'static,
 {
     #[allow(clippy::type_complexity)]
-    fn bob_spawn<AL: Ledger, BL: Ledger, AA: Asset, BA: Asset>(
+    fn spawn_alice<AL: Ledger, BL: Ledger, AA: Asset, BA: Asset>(
         &self,
         swap_request: rfc003::Request<AL, BL, AA, BA>,
         response: rfc003::Response<AL, BL>,
@@ -44,23 +43,20 @@ where
         S: Send + Sync + 'static,
     {
         let id = swap_request.id;
-        let seed = self.deps.seed.swap_seed(id);
+        let swap_seed = self.deps.seed.swap_seed(id);
 
         let (sender, receiver) = mpsc::unbounded();
 
         let swap_execution = {
             let ledger_events = self.deps.ledger_events.clone();
-            let state_store = Arc::clone(&self.deps.state_store);
 
             async move {
-                let bob_state = match response {
-                    Ok(accepted) => State::accepted(swap_request, accepted, seed),
-                    Err(declined) => State::declined(swap_request, declined, seed),
+                let alice_state = match response {
+                    Ok(accepted) => State::accepted(swap_request, accepted, swap_seed),
+                    Err(declined) => State::declined(swap_request, declined, swap_seed),
                 };
 
-                state_store.insert(id, bob_state.clone());
-
-                match bob_state {
+                match alice_state {
                     State {
                         swap_communication: SwapCommunication::Accepted { request, response },
                         ..
@@ -113,7 +109,7 @@ where
 
         let state_store = Arc::clone(&self.deps.state_store);
         tokio::spawn(receiver.for_each(move |update| {
-            state_store.update::<bob::State<AL, BL, AA, BA>>(&id, update);
+            state_store.update::<alice::State<AL, BL, AA, BA>>(&id, update);
             Ok(())
         }));
